Windows10DebloaterSilent
Changed all registry key code to include and perform a Test-Path, to check whether or not the path actually exists, and if not it will be created, along with the value specified. Beforehand the code *always* resolved to true. It will no longer only resolve to true.

I have made the Remove-Keys function, the Protect-Privacy function, and the Stop-EdgePDF function use the same $Debloat switch parameter as the Start-Debloat function.

I removed all forms of user interaction in the silent version of the script, and I have removed the Revert-Changes function since this is primarily to be used for silent debloating.

Credit to a60wattfish for the code on lines 154-158. Fantastic suggestion!