Added code to change reg keys/values in ntuser.dat
Changed all registry key code to include and perform a Test-Path, to check whether or not the path actually exists, and if not it will be created, along with the value specified. Beforehand the code *always* resolved to true. It will no longer only resolve to true.

I fixed a couple of bugs where I forgot the hyphen (-) before Verbose which lead to issues with -Verbose not working appropriately in some pieces of code.

I changed back the $_.name and $_.packagename portions of my Start-Debloat function from $_.AppxPackage and $_.AppxProvisionedPackage since they weren't uninstalling almost any bloatware, and they were uninstalling important Apps such as Store and Calculator (it was an amateur mistake).